+++ CHANGES.md
@@ -3,8 +3,12 @@ The released versions correspond to PyPi releases.
 
 ## Unreleased
 
+### Fixes
+* fixed condition parser to work with current DICOM standard
+
 ### Infrastructure
 * add Python 3.14 to CI (needs development version of `pydicom`)
+* update the tests for current version 2025b
 
 ## [Version 0.7.0](https://pypi.python.org/pypi/dicom-validator/0.7.0) (2025-05-25)
 Uses a grammar parser as condition parser.